Conflicts:
	Data/TestData.sqlite
	Source/SchemaProvider/SchemaProviderBase.cs

+++ Source/SchemaProvider/ParameterSchema.cs
@@ -9,7 +9,7 @@ namespace LinqToDB.SchemaProvider
 		public bool     IsIn          { get; set; }
 		public bool     IsOut         { get; set; }
 		public bool     IsResult      { get; set; }
-		public int?     Size          { get; set; }
+		public long?    Size          { get; set; }
 
 		public string   ParameterName { get; set; }
 		public string   ParameterType { get; set; }
